+++ internal/v4/export_test.go
@@ -6,7 +6,6 @@ package v4
 var (
 	BundleCharms                   = (*Handler).bundleCharms
 	ParseSearchParams              = parseSearchParams
-	StartTime                      = &startTime
 	DefaultIcon                    = defaultIcon
 	ArchiveCacheVersionedMaxAge    = &archiveCacheVersionedMaxAge
 	ArchiveCacheNonVersionedMaxAge = &archiveCacheNonVersionedMaxAge
